What kind of cast is this?

+++ cl_main.c
@@ -364,7 +364,7 @@ void CL_DNS_f (void) {
 	Q_strncpyz(address, Cmd_Argv(1), sizeof(address));
 	if (s = strchr(address, ':'))
 		*s = 0;
-	if (((int) addr.s_addr = inet_addr(address)) == INADDR_NONE) {
+	if ((addr.s_addr = inet_addr(address)) == INADDR_NONE) {
 		//forward lookup
 		if (!(h = gethostbyname(address))) {
 			Com_Printf("Couldn't resolve %s\n", address);
@@ -830,7 +830,7 @@ void CL_Init (void)
 
 	Key_Init ();
 	V_Init ();
-
+ 
 #ifdef __linux__
 	IN_Init ();
 #endif
